cleansing

+++ etl_web_to_gcs.py
@@ -20,7 +20,7 @@ def clean(df=pd.DataFrame)-> pd.DataFrame:
     df['passenger_count']=df['passenger_count'].astype('int64')
     df['payment_type']=df['payment_type'].astype('int64')
     df['RatecodeID']=df['RatecodeID'].astype('float64')
-
+    df['VendorID']=df['VendorID'].astype('int64')
     #df['lpep_pickup_datetime']=pd.to_datetime(df['lpep_pickup_datetime'])
     #df['lpep_dropoff_datetime']=pd.to_datetime(df['lpep_dropoff_datetime'])
     print(df.head(2))
